+++ packages/react-sdk/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.0.24](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-0.0.23...@stream-io/video-react-sdk-0.0.24) (2023-06-13)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.0.1`
+* `@stream-io/video-react-bindings` updated to version `0.0.11`
+
+### Features
+
+* add audio room demo app ([#572](https://github.com/GetStream/stream-video-js/issues/572)) ([77f7b65](https://github.com/GetStream/stream-video-js/commit/77f7b6596047d59e10c8e58abad38c4f48cc162f))
+
 ### [0.0.23](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-sdk-0.0.22...@stream-io/video-react-sdk-0.0.23) (2023-06-12)